Comentando codigo, aun no termino, continuo en casa

+++ app/resources/ArtistResource.php
@@ -6,6 +6,7 @@ use App\Models\Artist;
 
 class ArtistResource extends BaseResource implements IResource{
     
+    // Default params
     public function onConstruct(){
         $this->addParam("artist_name");
         $this->addParam("real_name");
@@ -13,24 +14,37 @@ class ArtistResource extends BaseResource implements IResource{
         $this->addParam("img_url");
     }
     
+    // Create new artist
     public function addAction() {
         $jsonRawBody = $this->request->getJsonRawBody();
         
-        $artist = $this->parse("Artist", $jsonRawBody);
+        $artist = $this->parse("Artist", $jsonRawBody); //Returns a phalcon model
         $artist->create();
         
         return $artist;
     }
-
+    
+    // Delete existing artist
     public function deleteAction($_identifier) {
         $artist = Artist::findFirst($_identifier);
+        
+        if(!$artist){
+            return ["msg" => "Artist does not exist"];
+        }
+        
         $artist->delete();
         
-        return "Artist deleted";
+        return ["msg" => "Artist deleted"];
     }
-
+    
+    // Get existing artist
     public function getAction($_identifier) {
         $artist = Artist::findFirst($_identifier);
+        
+        if(!$artist){
+            return ["msg" => "Artist not found"];
+        }
+        
         return $artist;
     }
